// src/services/synchronization/models/authentication-type.ts
/*---------------------------------------------------------------------------------------------
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/

export enum AuthenticationType {
  SERVICE = "Service",
  USER = "User",
}